--- src/Pages/Test/index.js
@@ -1,1 +1,0 @@
-export * from "./Test";
