+++ src/pages/Auth/AccountCreated/AccountCreated.js
