+++ src/pages/Auth/AccountCreated/index.js
