+++ src/pages/Dashboard/Wallet/index.js
