+++ src/pages/Dashboard/Wallet/molecules/RecieveQrCode/RecieveQrCode.js
