+++ src/pages/Dashboard/Wallet/molecules/RecieveQrCode/index.js
